+++ R/resp-url.R
@@ -2,7 +2,8 @@
 #'
 #' * `resp_url()` returns the complete url.
 #' * `resp_url_path()` returns the path component.
-#' * `resp_url_query()` returns the query component as a named list.
+#' * `resp_url_query()` returns a single query component.
+#' * `resp_url_queries()` returns the query component as a named list.
 #'
 #' @inheritParams resp_header
 #' @export
@@ -29,8 +30,19 @@ resp_url_path <- function(resp) {
 
 #' @export
 #' @rdname resp_url
-resp_url_query <- function(resp) {
+#' @param name Query parameter name.
+#' @param default Default value to use if query parameter doesn't exist.
+resp_url_query <- function(resp, name, default = NULL) {
+  check_response(resp)
+
+  resp_url_queries(resp)[[name]] %||% default
+}
+
+#' @export
+#' @rdname resp_url
+resp_url_queries <- function(resp) {
   check_response(resp)
 
   url_parse(resp$url)$query
 }
+